Storage format for collection ids increased from 16 to 32 bytes internally.
svn path=/trunk/eXist/; revision=6375


Former-commit-id: 5d7e5b7a06c9cac98d7f4edd949ec253f8c9a399

+++ src/org/exist/cluster/ClusterCollection.java
@@ -284,7 +284,7 @@ public final class ClusterCollection extends Collection {
         return collection.getDocumentCount();
     }
 
-    public short getId() {
+    public int getId() {
         return collection.getId();
     }
 
@@ -353,7 +353,7 @@ public final class ClusterCollection extends Collection {
         collection.store(txn, broker, info, node, privileged);
     }
 
-    public void setId(short id) {
+    public void setId(int id) {
         collection.setId(id);
     }
